test(ui): split indented Markdown cases below the file limit
